extracted basepresenterimpl from rxpresenter
