+++ app/src/main/java/lu/aqu/projper/mvp/BasePresenterImpl.java
@@ -1,0 +1,21 @@
+package lu.aqu.projper.mvp;
+
+public class BasePresenterImpl<T extends BaseView> implements BasePresenter<T> {
+
+    private T view;
+
+    @Override
+    public T getView() {
+        return view;
+    }
+
+    @Override
+    public void onViewAdded(T view) {
+        this.view = view;
+    }
+
+    @Override
+    public void onViewRemoved() {
+        view = null;
+    }
+}
