+++ app/src/main/java/lu/aqu/projper/mvp/RxPresenter.java
@@ -3,28 +3,17 @@ package lu.aqu.projper.mvp;
 import io.reactivex.ObservableTransformer;
 import io.reactivex.disposables.CompositeDisposable;
 
-public class RxPresenter<T extends BaseView> implements BasePresenter<T> {
+public class RxPresenter<T extends BaseView> extends BasePresenterImpl<T> {
 
     private CompositeDisposable disposer = new CompositeDisposable();
-    private T view;
 
     protected <D> ObservableTransformer<D, D> collectDispoable() {
         return upstream -> upstream.doOnSubscribe(disposable -> disposer.add(disposable));
     }
 
     @Override
-    public T getView() {
-        return view;
-    }
-
-    @Override
-    public void onViewAdded(T view) {
-        this.view = view;
-    }
-
-    @Override
     public void onViewRemoved() {
         disposer.clear();
-        view = null;
+        super.onViewRemoved();
     }
 }
